refactor: 엔티티별 커스텀 훅으로 분리 및 문자열 타입 enum화
