--- src/basic/constants/index.ts
@@ -1,9 +1,0 @@
-export * from "./notification";
-export * from "./search";
-export * from "./coupon";
-export * from "./discount";
-export * from "./cart";
-export * from "./product";
-export * from "./validation";
-export * from "./defaults";
-export * from "./calculation";
