+++ src/basic/constants/notification.ts
@@ -5,6 +5,3 @@ export const NOTIFICATION_TYPES = {
   SUCCESS: "success",
   WARNING: "warning",
 } as const;
-
-export type NotificationType =
-  (typeof NOTIFICATION_TYPES)[keyof typeof NOTIFICATION_TYPES];
