+++ src/basic/hooks/index.ts
@@ -1,1 +1,3 @@
 export * from "./useLocalStorage";
+export * from "./useCart";
+export * from "./useProducts";
